+++ unsubscribe_test.go
@@ -35,6 +35,9 @@ func TestUnsubscribe(t *testing.T) {
 	s.AddFilter("a/b/c")
 	s.AddFilter("d/e")
 
+	if v := s.Filters(); len(v) != 2 {
+		t.Error("expect 2 filters, got", v)
+	}	
 	if v := s.String(); !strings.Contains(v, "SUBSCRIBE --1-") {
 		t.Errorf("%q expect to contain %q", v, "SUBSCRIBE --1-")
 	}